refactoring_project/requirements.txt
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-WTF==1.0.0
gunicorn==20.1.0
h5py==2.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
MarkupSafe==2.0.1
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==9.0.0
six==1.16.0
Werkzeug==2.0.2
wincertstore==0.2
WTForms==3.0.1
tensorflow==2.7